@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "~14.18.0",
-    "npm": "~7.19.0"
+    "node": "~20.9.0",
+    "npm": "~10.1.0"
   },
   "main": "src/main.js",
   "scripts": {
