@@ -1,10 +1,12 @@
+import browser from 'webextension-polyfill'
+
 const basicNotification = {
   type: 'basic',
   iconUrl: 'icons/96.png',
   isClickable: true
 }
 
-module.exports.showPageAddedNotification = () => {
+export const showPageAddedNotification = () => {
   browser.notifications.create({
     title: 'Page added',
     message: 'Page added to your Sooner list',
@@ -12,7 +14,7 @@ module.exports.showPageAddedNotification = () => {
   })
 }
 
-module.exports.showPageAddFailedNotification = () => {
+export const showPageAddFailedNotification = () => {
   browser.notifications.create({
     title: "Couldn't add that page",
     message: 'Try using the toolbar button',
