@@ -9,7 +9,7 @@ precacheAndRoute(self.__WB_MANIFEST)
 const handler = createHandlerBoundToURL('/index.html')
 const navigationRoute = new NavigationRoute(handler, {
   denylist: [
-    new RegExp('/api/.*')
+    /\/api\/.*/
   ]
 })
 registerRoute(navigationRoute)
