@@ -1,5 +1,4 @@
 const { InjectManifest } = require('workbox-webpack-plugin')
-const CopyPlugin = require('copy-webpack-plugin')
 
 let config = {}
 
@@ -7,11 +6,6 @@ if (process.env.NODE_ENV === 'production' || process.env.VUE_APP_DEV_SERVICE_WOR
   config = {
     configureWebpack: {
       plugins: [
-        new CopyPlugin({
-          patterns: [
-            { from: 'public/', to: './' }
-          ]
-        }),
         new InjectManifest({
           swSrc: './src/service-worker.js'
         })
@@ -24,7 +18,7 @@ if (process.env.NODE_ENV === 'development') {
   config = {
     ...config,
     devServer: {
-      disableHostCheck: true,
+      allowedHosts: 'all',
       proxy: {
         '^/api': {
           target: 'http://localhost:3000',
