@@ -1,7 +1,8 @@
+import browser from 'webextension-polyfill'
+
 const KEY = 'HAS_READ_INTRO'
 
 export const hasReadIntro = async () => {
-  console.log(browser.storage)
   try {
     const results = await browser.storage.local.get(KEY)
     return results[KEY]
